# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Torneo de Tenis API",
        "contact": {
            "email": "julian.andres.arguello@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/tournaments": {
            "get": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Obtener lista de torneos",
                "description": "Devuelve la lista de torneos con sus géneros y estados asociados. \n     *                  Puedes filtrar los torneos por género utilizando el parámetro de consulta `gender` \n     *                  con valores `MALE` o `FEMALE`.",
                "operationId": "7810a369e52949739b11960d12c9f324",
                "parameters": [
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Filtro opcional por género de los torneos (MALE o FEMALE)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "MALE",
                                "FEMALE"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de torneos encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Estado de la operación",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Mensaje de la operación",
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "oneOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Tournament"
                                                    },
                                                    {
                                                        "$ref": "#/components/schemas/Tournament"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Crear y jugar un torneo",
                "description": "Crea un torneo con jugadores asignados, inicia el torneo y devuelve los resultados.",
                "operationId": "3bef609a811336de7bf6eb0a27b71652",
                "requestBody": {
                    "description": "Datos para crear un torneo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tournament_gender",
                                    "players"
                                ],
                                "properties": {
                                    "tournament_gender": {
                                        "description": "Género del torneo ('MALE' o 'FEMALE')",
                                        "type": "string"
                                    },
                                    "players": {
                                        "description": "Lista de jugadores para el torneo",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "name",
                                                "skill_level",
                                                "stength",
                                                "movement_speed"
                                            ],
                                            "properties": {
                                                "name": {
                                                    "description": "Nombre del jugador",
                                                    "type": "string"
                                                },
                                                "skill_level": {
                                                    "description": "Nivel de habilidad del jugador",
                                                    "type": "integer"
                                                },
                                                "stength": {
                                                    "description": "Fuerza del jugador",
                                                    "type": "integer"
                                                },
                                                "movement_speed": {
                                                    "description": "Velocidad de movimiento del jugador",
                                                    "type": "integer"
                                                },
                                                "reaction_time": {
                                                    "description": "Tiempo de reacción (opcional para jugadores femeninos)",
                                                    "type": "integer",
                                                    "nullable": true
                                                },
                                                "hooho": {
                                                    "description": "Atributo adicional opcional",
                                                    "type": "integer",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 2
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "maleTournament": {
                                    "summary": "Ejemplo de torneo masculino",
                                    "value": {
                                        "tournament_gender": "MALE",
                                        "players": [
                                            {
                                                "name": "Lucas Gonzalez",
                                                "skill_level": 70,
                                                "stength": 230,
                                                "movement_speed": 6
                                            },
                                            {
                                                "name": "Julián Ramirez",
                                                "skill_level": 68,
                                                "stength": 216,
                                                "movement_speed": 5
                                            },
                                            {
                                                "name": "Carlos Ramoz",
                                                "skill_level": 78,
                                                "stength": 228,
                                                "movement_speed": 8
                                            },
                                            {
                                                "name": "Juan Mirez",
                                                "skill_level": 56,
                                                "stength": 210,
                                                "movement_speed": 7
                                            }
                                        ]
                                    }
                                },
                                "femaleTournament": {
                                    "summary": "Ejemplo de torneo femenino",
                                    "value": {
                                        "tournament_gender": "FEMALE",
                                        "players": [
                                            {
                                                "name": "Maria Lopez",
                                                "skill_level": 88,
                                                "reaction_time": 900
                                            },
                                            {
                                                "name": "Ana Saucedo",
                                                "skill_level": 72,
                                                "reaction_time": 850
                                            },
                                            {
                                                "name": "Elina Perez",
                                                "skill_level": 68,
                                                "reaction_time": 260
                                            },
                                            {
                                                "name": "Jessica Bonnefon",
                                                "skill_level": 62,
                                                "reaction_time": 200
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Torneo jugado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Torneo jugado exitosamente"
                                        },
                                        "data": {
                                            "description": "Resultados del torneo",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error en los datos proporcionados"
                                        },
                                        "errors": {
                                            "description": "Detalles de los errores",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Gender": {
                "description": "Géneros",
                "properties": {
                    "id": {
                        "description": "ID único del género",
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "description": "Código del género",
                        "type": "string",
                        "example": "Male"
                    },
                    "title": {
                        "description": "Título del género",
                        "type": "string",
                        "example": "Masculino"
                    },
                    "created_at": {
                        "description": "Fecha de creación del género",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T16:31:13.000000Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización del género",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T16:31:13.000000Z"
                    },
                    "deleted_at": {
                        "description": "Fecha de eliminación del género",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Tournament": {
                "description": "Representación de un torneo",
                "properties": {
                    "id": {
                        "description": "ID único del torneo",
                        "type": "integer",
                        "example": 1
                    },
                    "gender_id": {
                        "description": "ID del género asociado al torneo",
                        "type": "integer",
                        "example": 1
                    },
                    "status_id": {
                        "description": "ID del estado del torneo",
                        "type": "integer",
                        "example": 3
                    },
                    "created_at": {
                        "description": "Fecha de creación del torneo",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T16:31:13.000000Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización del torneo",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T16:31:13.000000Z"
                    },
                    "deleted_at": {
                        "description": "Fecha de eliminación del torneo",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    },
                    "status": {
                        "$ref": "#/components/schemas/TournamentStatus"
                    },
                    "gender": {
                        "$ref": "#/components/schemas/Gender"
                    }
                },
                "type": "object"
            },
            "TournamentStatus": {
                "description": "Estado del torneo",
                "properties": {
                    "id": {
                        "description": "ID único del estado",
                        "type": "integer",
                        "example": 3
                    },
                    "code": {
                        "description": "Código del estado",
                        "type": "string",
                        "example": "Playable"
                    },
                    "title": {
                        "description": "Título del estado",
                        "type": "string",
                        "example": "Jugable"
                    },
                    "created_at": {
                        "description": "Fecha de creación del estado",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T16:31:13.000000Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización del estado",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T16:31:13.000000Z"
                    },
                    "deleted_at": {
                        "description": "Fecha de eliminación del estado",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}